Add requested changes from review
